ml_tracks/a_fire_danger/requirements.txt
absl-py==1.3.0
affine==2.3.0
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asciitree==0.3.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==21.2.0
backcall==0.2.0
basemap==1.3.6
basemap-data==1.3.2
basemap-data-hires==1.3.2
beautifulsoup4==4.10.0
bleach==6.0.0
blitz-bayesian-pytorch==0.2.8
#Bottleneck==1.3.2
cachetools==5.2.0
certifi==2021.10.8
cffi==1.15.0
cfgrib==0.9.9.1
cftime==1.5.1.1
charset-normalizer==2.0.7
click==8.0.3
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==2.0.0
comm==0.1.3
contourpy==1.0.6
cycler==0.11.0
dask==2021.11.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
eccodes==1.3.3
executing==1.2.0
fasteners==0.16.3
fastjsonschema==2.16.3
findlibs==0.0.2
#Fiona==1.8.22
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2021.11.0
geopandas==0.12.0
google-auth==2.14.1
google-auth-oauthlib==0.4.6
grpcio==1.51.1
h5netcdf==0.11.0
h5py==3.5.0
idna==3.3
importlib-metadata==5.1.0
importlib-resources==5.12.0
ipykernel==6.22.0
ipython==8.12.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.0.2
joblib==1.1.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
kiwisolver==1.3.2
locket==0.2.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.4.3
matplotlib-inline==0.1.6
mistune==2.0.5
multidict==6.0.2
munch==2.5.0
nbclassic==0.5.5
nbclient==0.7.3
nbconvert==7.3.0
nbformat==5.8.0
nest-asyncio==1.5.6
netCDF4==1.5.8
notebook==6.5.3
notebook_shim==0.2.2
numcodecs==0.9.1
#numpy==1.21.6
oauthlib==3.2.2
packaging==21.2
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.3
partd==1.2.0
pathlib==1.0.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pkgutil_resolve_name==1.3.10
platformdirs==3.2.0
pooch==1.5.2
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pydap==3.2.2
Pygments==2.14.0
pyparsing==2.4.7
#pyproj==3.4.0
pyrsistent==0.19.3
pyshp==2.3.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2021.3
#PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.2
QtPy==2.3.1
rasterio==1.2.10
requests==2.26.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rioxarray==0.12.2
rsa==4.9
scikit-learn==1.0.1
scipy==1.7.2
seaborn==0.11.2
Send2Trash==1.8.0
Shapely==1.8.5.post1
six==1.16.0
sniffio==1.3.0
snuggs==1.4.7
soupsieve==2.3
stack-data==0.6.2
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.17.1
threadpoolctl==3.0.0
tinycss2==1.2.1
toolz==0.11.2
#torch==1.12.1+cu113
#torchaudio==0.12.1+cu113
#torchvision==0.13.1+cu113
tornado==6.2
tqdm==4.62.3
traitlets==5.9.0
typing-extensions==3.10.0.2
uri-template==1.2.0
urllib3==1.26.7
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
WebOb==1.8.7
websocket-client==1.5.1
Werkzeug==2.2.2
widgetsnbextension==4.0.7
xarray==0.20.1
xbatcher==0.2.0
xgboost==2.1.4
yarl==1.8.1
zarr==2.10.2
zipp==3.11.0
